feat: Impl. exception for errors 404 and 500.

+++ src/main/java/com/dev/praticas_spring/service/ProdutoService.java
@@ -1,10 +1,10 @@
 package com.dev.praticas_spring.service;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
+import com.dev.praticas_spring.exceptions.RecursoNaoEncontradoExeception;
 import com.dev.praticas_spring.model.Produto;
 import com.dev.praticas_spring.repository.ProdutoRepository;
 
@@ -25,11 +25,16 @@ public class ProdutoService {
         return produtoRepository.findAll();
     }
 
-    public Optional<Produto> buscarPorId(Long id){
-        return produtoRepository.findById(id);
+    public Produto buscarPorId(Long id){
+        return produtoRepository.findById(id).orElseThrow(() -> new RecursoNaoEncontradoExeception("Produto com ID (" + id + ") não encontrado"));
     }
 
     public void deletarProduto(Long id){
+
+        if (!produtoRepository.existsById(id)){
+            throw new RecursoNaoEncontradoExeception("Produto com ID (" + id + ") não encontrado");
+        }
+        
         produtoRepository.deleteById(id);
     }
 }
